refactor: 테스트를 위한 비밀 번호 암호화 객체 추상화 [#38]
