+++ src/main/java/com/greedy/zupzup/auth/infrastructure/BCryptPasswordEncoder.java
@@ -1,0 +1,18 @@
+package com.greedy.zupzup.auth.infrastructure;
+
+import org.mindrot.jbcrypt.BCrypt;
+import org.springframework.stereotype.Component;
+
+@Component
+public class BCryptPasswordEncoder implements PasswordEncoder {
+
+    @Override
+    public String encode(String rawPassword) {
+        return BCrypt.hashpw(rawPassword, BCrypt.gensalt());
+    }
+
+    @Override
+    public boolean matches(String rawPassword, String encodedPassword) {
+        return BCrypt.checkpw(rawPassword, encodedPassword);
+    }
+}
