+++ src/main/java/com/greedy/zupzup/auth/infrastructure/PasswordEncoder.java
@@ -1,0 +1,6 @@
+package com.greedy.zupzup.auth.infrastructure;
+
+public interface PasswordEncoder {
+    String encode(String rawPassword);
+    boolean matches(String rawPassword, String encodedPassword);
+}
